selftests: Refactor the run-perf selftests
There is quite an overlap in code, let's simplify and share the code.
